backend/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
pymysql==1.0.2
retrying==1.3.3